+++ Modules/People/Resources/lang/de/people.php
@@ -197,4 +197,6 @@ return [
     'column_set' => 'Spalten',
     'name_nationality_occupation' => 'Nationalität, Beschäftigung',
     'contact_info' => 'Kontaktinformationen',
+    'has_tax_number' => 'Hat eine Steuernummer',
+    'applied' => 'Hat sich beworben',
 ];